order aggregate gets order number from sequence provider, not from command

+++ Shop.Domain/Aggregates/OrderAggregate/Events/OrderCreated.cs
@@ -6,12 +6,12 @@ namespace Shop.Domain.Aggregates.OrderAggregate.Events
 {
     public class OrderCreated : DomainEvent
     {
-        public int Number { get; }
+        public long Number { get; }
         public Guid User { get; }
 
         public OrderStatus Status { get; }
 
-        public OrderCreated(Guid sourceId, int number, Guid user, OrderStatus status = OrderStatus.Created):base(sourceId)
+        public OrderCreated(Guid sourceId, long number, Guid user, OrderStatus status = OrderStatus.Created):base(sourceId)
         {
             Status = status;
             User = user;
